backend/requirements.txt

Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
channels==4.1.0
channels-redis==4.2.0
django-cors-headers==4.6.0
Pillow==11.1.0
scikit-learn==1.7.1
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
zope.interface==7.2